Refactor user and trip models to use uint64 for IDs; update related services and repositories
- Changed ID and UserID fields in TripResponse, UserResponse, Trips, and User models from int to uint64.
- Updated CreateTrip and RegisterUser methods to accept uint64 for user IDs.
- Refactored user service to utilize a repository pattern for database operations.
- Implemented error handling improvements with custom error types for better clarity.
- Added a new error handling middleware for centralized error responses.
- Introduced a response package for standardized success responses.

+++ go-service-driver/internal/dto/trip_request.go
@@ -15,8 +15,8 @@ type CreateTripRequest struct {
 }
 
 type TripResponse struct {
-	ID                 int        `json:"id"`
-	UserID             int        `json:"user_id"`
+	ID                 uint64     `json:"id"`
+	UserID             uint64     `json:"user_id"`
 	Title              string     `json:"title"`
 	Description        string     `json:"description"`
 	StartDate          *time.Time `json:"start_date"`
